refactor: extract Public Sharing into a dedicated module
- Add server/utils/public-sharing.ts owning the Public Sharing lifecycle:
  enablePublicSharing (eligibility + token rotation + locking/retry),
  disablePublicSharing (token invalidation), and resolvePublicIssue (token ->
  Public Issue Page with the uniform "not found" and masked headers)
- Split protected-issue.ts (Protected Issue View) out of the old public-issue.ts
  so each file matches one CONTEXT.md concept; delete public-issue.ts
- Slim share.post/share.delete and the public token route to thin orchestrators
  (-138 lines); the public-read invariants now live behind one seam
- Rename public-issue.test.ts -> public-sharing.test.ts and point the protected
  issue routes at the new module path
