+++ server/api/projects/[projectId]/issues/[issueId].get.ts
@@ -1,5 +1,5 @@
 import { getAccessibleIssue } from '~~/server/utils/issue-access'
-import { buildProtectedIssueResponse } from '~~/server/utils/public-issue'
+import { buildProtectedIssueResponse } from '~~/server/utils/protected-issue'
 
 export default defineEventHandler(async (event) => {
   const { db, project, issue } = await getAccessibleIssue(event)
